requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
Pillow>=10.0
requests>=2.31
channels>=4.0
channels-redis>=4.1
daphne>=4.0
python-decouple>=3.8
